create separate AddCardButton file
